Fix further tests
